add player and light switch for big scene

+++ Assets/Client/Model/Position.cs
@@ -4,9 +4,9 @@
     {
         public string clientId { get; set; }
 
-        public int x { get; set; }
+        public float x { get; set; }
 
-        public int y { get; set; }
+        public float y { get; set; }
     }
